Start moving HeatMap to TS
